examples/django_app/requirements.txt
Django==1.11.7
gunicorn==19.7.1
djangorestframework==3.7.3
psycopg2==2.7.3.2
requests==2.18.4
pytest==3.2.3
bg-kube==0.0.6
